Move middlewares to mw file 🚚
